// src/types.ts
export interface Subject {
  id: string;
  name: string;
  color: string;
  isArchived?: boolean;
}